+++ spec/dnsimple/client/contacts_spec.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 
 describe Dnsimple::Client, ".contacts" do
 
-  subject { described_class.new(api_endpoint: "https://api.dnsimple.test", access_token: "a1b2c3").contacts }
+  subject { described_class.new(base_url: "https://api.dnsimple.test", access_token: "a1b2c3").contacts }
 
 
   describe "#contacts" do