+++ spec/dnsimple/client/zones_records_spec.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 
 describe Dnsimple::Client, ".zones" do
 
-  subject { described_class.new(api_endpoint: "https://api.dnsimple.test", access_token: "a1b2c3").zones }
+  subject { described_class.new(base_url: "https://api.dnsimple.test", access_token: "a1b2c3").zones }
 
 
   describe "#records" do